extract config helper to seperate file
